% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetview_palette.R
\name{get_palette}
\alias{get_palette}
\title{Generate a named color palette for discrete classes}
\usage{
get_palette(classes, others_label = "Others")
}
\arguments{
\item{classes}{Character string.
The discrete class names or factor levels to map to colors.}

\item{others_label}{Character, (default = "Others").}
}
\value{
A named character vector where names correspond to
}
\description{
Generate a named color palette for discrete classes
}
\examples{
NULL

}
