% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetview_palette.R
\name{scale_fill_ggnetview}
\alias{scale_fill_ggnetview}
\title{Custom discrete fill scale for ggNetView}
\usage{
scale_fill_ggnetview(
  classes,
  ...,
  others_label = "Others",
  na_value = "#e0e0e0",
  drop = FALSE
)
}
\arguments{
\item{classes}{Character string.
The discrete class names or factor levels to map to colors.}

\item{...}{Additional arguments passed to `ggplot2::scale_color_manual()`.}

\item{others_label}{Character, (default = "Others").}

\item{na_value}{Color for missing values. Default `"#e0e0e0"`.}

\item{drop}{Logical, passed to `ggplot2::scale_color_manual()`.}
}
\value{
A `ggplot2` scale object.
}
\description{
Custom discrete fill scale for ggNetView
}
\examples{
NULL
}
