% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_TOM_in_WGCNA.R
\name{trans_TOM_in_WGCNA}
\alias{trans_TOM_in_WGCNA}
\title{Translate TOM matrin to create graph object}
\usage{
trans_TOM_in_WGCNA(TOM, mat, threshold = NULL)
}
\arguments{
\item{TOM}{matrix
TOM matrix from WGCNA result}

\item{mat}{matrix
matrox to WGCNA analysis}

\item{threshold}{numeric
the threshold in weight}
}
\value{
A Data frame contain from, to and weight
}
\description{
Translate TOM matrin to create graph object
}
\examples{
NULL
}
